requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
google.cloud==0.34.0
google.oauth2==2.30.0
pydub==0.25.1